[FIX] social_*: allow to upload image in comment
Bug
===
Since odoo/odoo@e8567dadbdaaec6e5be8f66f12e2b02da321f999 , it's not possible to
upload images in comments, in social.

Now, the file data is supposed to be managed in the onChange event,
and we can not rely on the input value.

Task-3339389

X-original-commit: db93ad79ce5b26e69d8f810a82bffb5a580e7395
Part-of: odoo/enterprise#46743