[FIX] sale_subscription: hide payment method button in portal
https://github.com/odoo/enterprise/pull/37536 introduced the subscription_state
field and mistakenly made the payment method button visible for closed
subscriptions. This commit hides the button.

A subscription can be reopened by making a payment through the payment form. The
payment method used is automatically assigned to the subscription, so there is
no need to set or manage the payment method while a subscription is still
closed.

task-3451159

closes odoo/enterprise#45594

X-original-commit: 7736c6aa9b24f4e64342c4862cb8e508c52863e3
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>