[FIX] account_disallowed_expenses_fleet: test data
The aim of this commit is to fix the tax testing data in order to be
able to enforce the constrains again.

Context:
The `_validate_repartition_lines` isn't run correctly since the merging of
`invoice_repartition_line_ids` and `refund_repartition_line_ids` into
the same field.

Before the commit:
The test doesn't pass with the constrains enforced.

After the commit:
The test pass with the constrains enforced.

closes odoo/enterprise#45215

Task-id: None
X-original-commit: 81d4ef313e095765f2c6a1252be34cf331408612
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>