[FIX] web_studio: command palette should open within studio
Be inside studio, on the HomeMenu.
Open the command palette and click on an item to open the corresponding action.

Before this commit, the action opened outside studio. This was because the command provider for menus
triggered the selectMenu method of the menuService directly which bypasses the studio mechanism.
This had a huge side effect: studio was never really exited, meaning that the user context was dirty.

After this commit, the studio service does the necessary override to redirect the selection of a menu
to within studio.

closes odoo/enterprise#45293

X-original-commit: a2079e704ae7e7c93f252a9a188165f927b3dfd4
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>