[FIX] mrp_mps: get the right quantities in Master production Schedule
Current behaviour:
In the Manufacturing module, we get inconsistant values depending on the expected date and time
of purchase orders

Steps to reproduce:
- Set 'Time range' of Master production Schedule to 'Daily' in Manufacturing's settings
- Create a product and his BOM.
- Create a purchase order on that product with an expected date of '2023-08-04 00:30:00', quantity: 10 and confirm.
- Create another purchase order on the same product with an expected date of '2023-08-04 11:30:00', quantity: 20 and confirm.
- Go to Master Production Schedule in Manufacturing
- Add the created product to be displayed.

Issue:
On the 'Suggested Replenishment' we have quantity 10 on 'Aug 3' and 20 on 'Aug 4'.

Expected:
We must have 30 on 'Aug 04'

Reason of the problem:
We call search on ``StockMove`` and ``PurchaseOrderLine`` in a range date which doesn't consider the time, and the timezone of the user.

Solution:
Create a range date in utc timezone specially for search, retreive the records and map the quantities to dates with the timezone.

opw-3299645

closes odoo/enterprise#45174

Signed-off-by: William Henrotin (whe) <whe@odoo.com>