[FIX] sale_subscription: correct renewal state for parent order
The renewal state of parent is ambiguous when the renewal order is sent as quotation. The state of parent would change to recurring when the renewal is still in quotation sent stage.

1. Create a recurring SO.

2. Create a renewal.

3. Change the state of renewal and go back to check the status of parent order.

-The renewal is in quotation -> The renew_state of the parent SO is renewing

-The renewal is in quotation sent -> The renew_state of the parent SO is N/A

-The renewal SO is confirmed -> The renew_state of the parent SO is renewed

Current Behavior:
The status of parent SO is recurring.

Expected Behavior:
When the renewal is in quotation sent, the state of the parent SO should be renewing

OPW-3444959

closes odoo/enterprise#45243

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>