[IMP] l10n_{cl,co,ec,pe}_edi: Disable account_edi auto_install
We change account_edi to be auto_install=False to avoid installing it when
it is not nedeed and create noise data that will always be empty.
It was previously auto_installed when Invoicing was installed.

Note that auto_install=['parent_module'] will automatically install the module
when 'parent_module' is installed (in this case the localization) AND install
missing dependencies (in particular here 'account_edi').

Community: https://github.com/odoo/odoo/pull/130836
Task-3454076

closes odoo/enterprise#45337

X-original-commit: 853724eddd6eee17deb8929d02c8baf4c14defb1
Related: odoo/odoo#130997
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>