[FIX] currency_rate_live: update log while unable connect online exchange rate
This issue will be generated when retrieving real-time currency rate updates
if the connection cannot be established or if the web services are unavailable
at that time.

Error message :
Unable to connect to the online exchange rate platform bbr. The web service may
be temporary down.

The logger is updated to use the 'warning' level instead of the 'exception'
level. This change reflects a less severe logging level for cases when  unable
to connect online exchange rate

sentry-4330375314

closes odoo/enterprise#46053

X-original-commit: 989bfdc606b2f2f0f90cba24395bbe37672726af
Signed-off-by: Laurent Smet (las) <las@odoo.com>