[FIX] sign: make sign template PDF readonly
The sign.template datas field is writable, which makes it very easy to replace
or delete the PDF. Modifying the underlying document doesn't make sense and
causes integrity checks to fail. This commit prevents these issues by making the
field readonly. Consequently, the Edit and Clear buttons in Template Properties
are no longer visible.

Note that the document itself can still be modified (for example, via the
Documents app), causing integrity checks to fail. However, that is a separate
topic.

task-3455091

closes odoo/enterprise#45269

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>