[FIX] sale_planning: ensure to update records in self in compute
Before this commit, the compute alters the records received by the
`_read_group` called inside that compute (`_compute_planning_hours`).
The problem, by doing that, the orm will directly do a write to update
the cache and the db since the records receive by `_read_group` are
not new records as it is the case in the self when the compute is
triggered by the onchange method.

This commit ensures the records altered inside the compute are the
ones contained in the self to avoid trigger the write method during
the onchange method.

closes odoo/enterprise#45329

X-original-commit: c39f6d3075d1481284b1f6f5f428a2919975040c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>