[FIX] helpdesk: improve btn-primary text contrast
=== ISSUE ===

If you navigate to the Helpdesk module dashboard, there is a banner with
statistics and state of tickets. Some of these are primary buttons made
to access your tickets faster.

These `btn-primary` include a `<h2>` tag to make it fill all the available
space but this structure also means you need to set a `text-white` to the
title to make it white as the other elements inside the button, but doing
so, it is not handled when you hover the button etc.

=== AFTER ===

We fix the structure of the `btn-primary` element to make it easier to
control and handle the color of the `<h2>` tag which we turned into a
`<span>` element with the same styling as before.

task-3455273
part of task-3326263

closes odoo/enterprise#45462

X-original-commit: 6afcb644be1418e807b1fa709e3347210e587ab4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>