[FIX] sign: fix after relational model changes
After https://github.com/odoo/enterprise/pull/43125/ was merged, the x2many
records that are rendered in the sign template edition are not being shown
properly (e.g. tags in the top bar, selection options in the popover of a
selection field). This commit fixes these issues by changing how we set the
dependent fields of fields that use the relational model.

task-3455387

closes odoo/enterprise#45295

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>