[IMP] account_reports: generalize activity plan / scheduler
By introducing a new wizard that allows to create activity and apply plan on
multiple records at once, we have introduced a new (UI) way to create activity.
We apply here the same changes done on the activity form that was formerly used
to create activity to the wizard that is now in charge of it.

See odoo/odoo#134007

Task-3390865

Part-of: odoo/enterprise#48599