[FIX] timesheet_grid: start timer in project without company set
Before this commit, when the user starts a timer on a timesheet
linked to a project without company set, the timer header will display
the following message: "A timer is running in another company.".
This message is not correct since the project has no company and so
the project is visible for all companies.

This commit fixes the issue by checking first if the project's company
is set before checking if the company is in the current ones. By
doing that, the user can now edit the timesheet data in timesheet
timer header component.

closes odoo/enterprise#45327

X-original-commit: d332388ee16b4c24a4a8d77b8fe9e209c9ab5b51
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>