[FIX] account_reports : show correct period on clicking amount PL report
Issue:
======
transaction of this months always appears on clicking on other months
amount in P&L report for other months when we have this month filter on
and comparaison on.

Steps to reproduce the error:
=============================
- Go to accounting/reporting/Profit and Loss
- Select filter as this month
- select comparaison and add 3 in 'Number of periods' of 'Pervious period'
- Click on amount for some previous month and it will shows transaction
  for this month

Solution:
=========
The value of the key 'column_group_key' was wrong on the call from the
js side , it should be the value on the cell and not the value of the
first column of the line.

opw-3440776
opw-3423204

closes odoo/enterprise#45448

X-original-commit: fba99f3e92b39020dcd86ac8e505627a1867910e
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>