[FIX] sale_amazon: use current company location when warehouse not have location
The Error is generated when the `_sync_pickings()` method is executed
for the cron job` Amazon: sync delivery orders`. It is because the value of
partner_id (location_) here is False.

Error: 'bool' object is not subscriptable  while evaluating
'model._sync_pickings()'

link to the PR from where the issue is generated:
https://github.com/odoo/enterprise/pull/29378/files#diff-36fff3c417ab5ebfcd2ce305afdd6c84ba8af4c577c1475a581effda06764272R236

This commit solves the issue by providing the current company location when
the warehouse does not have a location.

sentry-4369020199

closes odoo/enterprise#46504

X-original-commit: 2c88a8264a1054f7cf36abfdf3a95bc3f60d7223
Signed-off-by: Achraf Ben Azzouz (abz) <abz@odoo.com>