[FIX] account_followup: select only sms tmpl w/ `res.partner` model
Issue:
On account follow up we can set an SMS template whose model is not
`res.partner`, and if in said template we try to access a field that
is not of `res.partner`, we have an error when rendering the
template when triggering a followup. We should prevent users from
selecting templates that are not based on `res.partner`.

Steps to reproduce:
- Install Accounting, SMS and another app (FSM for example)
- Accounting > Configuration > Follow up levels > 15 days
- Check SMS, set a template on it, on from FSM (so the model of the
  template is not `res.partner`).
- In the body add a reference to an attribute that is not present in
  `res.partner` for example {{object.origin}}
- Accounting > Customers > Follow-up Reports
- Select one of the entry that use the follow-up we modified
- Click on "Follow up", we are met with an error of failing to
  render the template.

Cause:
The Follow-up expects the template to be based on `res.partner`, but
it's not the case always, because the user can set any sms template
they want, regardless of the model. So when we access an attribute
that doesn't exist on said model, we have a rendering failure of the
template.

Fix:
Add a domain on `sms_template_id` on `FollowupLine` to force the
model to be only `res.partner`. This is already the case in
`AccountFollowupReport`. Also added said domain on the mail_template_id,
following the same logic.

Affected versions:
16 up to master

Reference:
opw-3372220

closes odoo/enterprise#45359

Signed-off-by: Cedric Snauwaert <csn@odoo.com>