[FIX] currency_rate_live: uses rate valid for today
Steps to reproduce:
- choose the FTA currency rate
- run manually

Issue:
The rates will be the ones of tomorrow which is not really an issue per se (as we will use the one before for Today's transactions) but can be counterintuitive for the use.

Solution:
We are fetching the rate of Yesterday.

If we are on Monday the 07th of August 2023, the one of Yesterday will be the one released on Friday the 4th of August (valid for following Saturday, Sunday and Monday).
It is more intuitive for the user to have the "Last Update Date" rate valid for Today's update.

Note:
FTA rate released on Monday 7th will ba valid for Tuesday 8th and
"If no exchange rates were entered for the specified date (weekend, public holidays), the rates entered for the last working day before the specified date will be shown."
https://www.rates.bazg.admin.ch/home

opw-3379010

closes odoo/enterprise#45391

X-original-commit: 57a87d79171aaeb32be32a681a62f1e4c656fa7a
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>