[IMP] web_gantt: highlight current month in Gantt view
Currently, when the scale of the Gantt view is set to 'week' or 'month',
the column corresponding to the current day is highlighted. This commit
extends this behavior to highlight the current month when the scale of
the Gantt view is set to 'year' or the current hour when the scale is
set to 'day'.

task-3458164

closes odoo/enterprise#45415

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>