[REF] : t-att- should be t-att=*
As all the templates are now imported in the owl app, the templates must
comply to owl.

Some attributes are treated specially by Owl: it does not set the
attribute, but only the corresponding DOM property (eg, t-att-value will
set the value property, ie el.value will have the expected value, but
el.getAttribute("value") will be null). To force Owl to set the
attribute instead of treating it as a property, this commit converts
some compute attributes from the t-att-* form to the t-att="{*: }" form,
which bypasses Owl's special handling from them.

Part of task~3443861

Part-of: odoo/enterprise#45020