[FIX] account_reports: sort lines flickering
When resetting the sort lines to the default order, there was a flickering.

This was due to the fact that the previous 'lines_order' was not properly erase from the data, so it was briefly applied before being recalculated.

To see the flickering you need to do a full 'sorting cycle' which means that you have to sort 'asc', 'desc' and 'default' and then sort once more.

closes odoo/enterprise#45517

X-original-commit: 84c34bb2a5349a5a07fede4dc969a805f52cf732
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>