[FIX] planning: fix crash when no "shift_planned"
__Current behavior before commit:__
When the filter "shift_planned" does not exist (like on odoo.com), the
page crashes when a filter is removed from the search bar.
The page also crashes when pressing the "Auto Plan" button.

__Description of the fix:__
Check if `this._getHighlightPlannedSearchItems()` is not null before
retrieving its attributes.

__To reproduce:__
On runbot, first remove the "shift_planned" filter in
"planning.slot.search" view at line 18.

On odoo.com, just go to the Planning app and remove the default filter.

opw-3441371

closes odoo/enterprise#45507

X-original-commit: 4a57e2cfc2650ca2c32ba6c23f3bcbba5f53fc89
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>