[FIX] *: external layout DIN5008 for followup report and task report
*: l10n_din5008_account_followup,l10n_din5008_industry_fsm

The DIN5008 layout is mandatory in Germany, Austria and a part of Switzerland.
The community module l10n_din5008 implements it but it is currently
still not working correctly for (at least) the follow-up report (from account_followup)
and the task report (from industry_fsm).

This commit makes the DIN5008 layout work for both of the
aforementioned reports.
The new modules 'l10n_din5008_account_followup' and
'l10n_din5008_industry_fsm' are added as part of this fix.

task-3394263

closes odoo/enterprise#44908

Related: odoo/odoo#130337
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>