[FIX] l10n_es_reports: fix exclusion of partners from of MOD347
Steps to reproduce:
- Install Accounting and configure Spanish localization
- Create a contact (e.g. Contact X)
- Create 4 invoices for Contact X and confirm them:
  * Amount: 1000€ - Invoice Date: 19/01/2022
  * Amount: 1000€ - Invoice Date: 19/04/2022
  * Amount: 1000€ - Invoice Date: 19/07/2022
  * Amount: 1000€ - Invoice Date: 19/10/2022
- Go to Accounting / Reporting / Statement Reports / Tax Report
- Select Report: Tax Report (Mod 347) (ES)
- Select curtom dates from 01/01/2022 to 31/03/2022
- Check line "B - Entregas de bienes y prestaciones de servicios superiores a 3.005,06 €"
The line is empty.
As the sum of all invoices of Contact X for the fiscal year 2022 is higher than 3005,06€ (i.e. 4000€),
the amount of all invoices of Contact X for the filtered period should appear (i.e. 1000€).

Cause:
It is due to the fact that partners are excluded from the report if the sum of their invoices
doesn't reach the threshold of 3005,06€ in the selected period, when the check to exclude
partners should be based on the whole fiscal year.

Solution:
When computing the partners to exclude, the fiscal year is retrieved from the start date
(selected in the report) and the computation is based on that fiscal year.
In the particular case where the start and the end dates are from different fiscal years,
nothing will be done (computation based on the selected dates).

opw-3357320

closes odoo/enterprise#45439

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>