[FIX] account_accountant: Don't get successful reconciliation log when it's not the case
Steps to reproduce:
- Enable batch payment in accounting settings.
- Go to Accounting > Vendors > Payments > Select Multiple, but at least 2 records need to
 be different vendors, create a batch and Validate.
- Go to Dashboard > Bank > New > Create a new transaction that matches one of the values in
the batch payment, as well as making sure to set the Label field for this transaction to be
exactly the Reference/Memo from the record you've chosen from the batch payment.
- Go to Configuration > Reconciliation Models > Invoices/Bills Perfect Match > Run Manually.
- Return to Bank Journal from Accounting Dashboard > Discuss Log of the added transaction.

Issue:
Note created which says ”This bank transaction has been automatically validated using the reconciliation model”

Reason of the problem:
We don't check if the statement line is reconciled before logging the success message.

opw-3419358

closes odoo/enterprise#45431

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>