[FIX] l10n_de_pos_cert: ensure VAT rates are loaded
Steps to reproduce:
1. set up `l10n_de_pos_cert` and make a PoS in German company
2. make an order, make the payment and validate
3. add another order line, do not pay it and close the session
4. re-open the session, the PoS will never load

A tax error is raised and blocks the PoS from loading.

The issue happens because when checking for valid VAT rates,
`vatRateMapping` variable is not initialized. To fix this
initialization of orders from local storage should await the result of
`vatRateMapping` from fiskaly so the check can work.

opw-3439224

closes odoo/enterprise#45500

X-original-commit: d6ac102a1e0ade2ae59ed783a7fc7f6ac3b3bbbb
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>