[FIX] sign: bad comparison operator made sign reminders spammy
A simple typo meant that email reminders would get sent every day no
matter the configuration of the request.

This commit also slighlty improves the readability of the query and a
comnment to clarify why it is built the way it is (with an OR, which
is a bit suprising).

OPW-3455341

closes odoo/enterprise#45533

X-original-commit: d7fed869d7dfee4296cee99c3560a4452410bc44
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>