[FIX] spreadsheet_edition: ignore list json fields
Go to Accounting > Accounting > Journal Items
Export the list view to a spreadsheet (Favorites > insert list in spreadsheet).
=> the column "Analytic" is filled with errors "Fields of type
"json" are not supported" [^1].

With this commit, the column isn't even exported to the spreadsheet.
since it can only bring confusion to the user.

It's probably not the user's intention to export the json field anyway.

Discussed with AL.

[^1]: see commit odoo/odoo@9e6d598e6a96434c68634149ecf1acf3e714ca03

closes odoo/enterprise#45476

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>