[IMP] account_reports: Showing invoice date on reports
Description of the issue/feature this PR addresses:

The invoice date (or accounting date if there is not accounting date available)
is not shown in the General Ledger, Partner Ledger and Journal Report.
This piece of information is missing and could be useful

During this feature development, a bug was found in the Customer Invoices in
the Journal Report. The name of the account receivable of the invoice was
renamed from [code][name] to [code][partner] to give more information
to the user. The bug was that this behavior was only present in the first line
of the invoice so if the invoice was paid in multiple times, only the first
line would have the adapted name.

Desired behavior after the PR is merged:

Adding this commit, the three targeted reports now show the invoice date and
a small fix is done to the account name in the Journal Report.

The General Ledger now has an "Invoice Date" column which shows the data
resulting from retriving both the accounting and the invoice date and showing
the latter only if the former is not available.

The Partner Ledger now uses the old "Ref" column as column name. That column
is removed and replaced by a new "Invoice Date" name which content is the
date previously used in the column's name.

The Journal Report now has a new "Invoice Date" column and doesn't use
the date in the name of the column anymore.

With regards to the bug that was found, a fix is made so that every line
showing a payment in the Customer Invoices of the Journal Report uses
the adapted name [code][partner].

task-3388333

closes odoo/enterprise#46498

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>