[FIX] l10n_mx_edi: export mass edi documents
When selecting several invoices from the list view, it is possible to trigger an action to export all edi documents in a zip file. In 16.4, it was no longer possible to download edi documents for the Mexican localization because they are stored in a different field. With this commit, we extend the function that provide all edi documents to export to also include Mexican ones.

task-130177

closes odoo/enterprise#45529

Related: odoo/odoo#131242
Related: odoo/upgrade#5032
Signed-off-by: Laurent Smet (las) <las@odoo.com>