[FIX] l10n_cl_edi : fix creation of duplicate vendor bills
This is basically applying the same logic as this [commit](https://github.com/odoo/enterprise/pull/36821)
but in a slightly different context.

To reproduce:
- create a chilean localization database with a configure DTE incoming
email server to fetch valid EDI XML from a email server (in this case
a valid vendor bill)
- misconfigure the default account "Proveedores" for purchase balacing,
by changing the type from "Payable" to another type of liability, i.e.
 "Current Liabilities"
- try to fetch the XML (`fetch_mail`)

Behavior before fix:
- duplicated bills with same name will show up in vendor bills
- first bill is invalid but has fields prefilled and committed (invoice
lines, customer, due date, etc.)
- the second bill is a bare-bones invoice with all error encountered during
the XML handling and inviting the user to manually check the XML (chatter
attachment) and fill out the information manually

Behavior after fix:
Only one vendor bill gets created (bare-bones invoice inviting user
to correct it manually)

Context:
While fetching the emails from the “DTE” server and parsing the attachment,
while handling a valid xml (customer example attached) the code will run
the `_create_document_from_attachment` function.

It will hit this try block :
https://github.com/odoo/enterprise/blob/16.0/l10n_cl_edi/models/fetchmail_server.py#L262

The logic seems to be that we will try to generate a invoice in “Form” mode
(pre-16 context), that now is supposed to be equivalent to the with context here:
https://github.com/odoo/enterprise/blob/16.0/l10n_cl_edi/models/fetchmail_server.py#L319
If this for some reason fails, the except block is supposed to capture
the error, create a new bare-bones invoice and simply log all the issues
in the chatter, thus inviting the user to check manually why this vendor
bill as an issue and correct it manually.

While doing so, the logic tries to fetch all relevant info from the XML
fields, fill in the field and then “commit” it using the invoice_form.write() call.

The issue is that the customer has configured his purchase journal to only
use one account the is set to “Current Liabilities” instead of “Payable”.
This will hit this constrains check `_check_payable_receivable`:
https://github.com/odoo/odoo/blob/16.0/addons/account/models/account_move_line.py#L1204

So in a normal setting, when using the frontend, trying to submit a vendor
bill with theses invoice lines, would have triggered the UserError and
prevent the user from committing that bill.
But since these checks only happened at the function return of
`_get_invoice_form`, it’s too late since the invalid invoice was already
committed and there is no rollback.

What happens next is that we hit the except block, we capture the error
message and then a new fresh invoice form is created with barebone information.
The ideas is to validate the rest of the information in the XML, log any
issues in the chatter, add the XML as attachment and let the client decide
what to do with this.

So de facto, right now two invoice with the same name are create, one
having invoice lines and journal entries that are not valid. Then a
second one that is supposed to be the catchall invoice to be manually
fixed by the client.

So the proposed fix just forces a rollback if the first invoice to be
committed raises an error, so we only keep the second invoice created
during the except block.

opw-3375744

closes odoo/enterprise#45535

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>