[IMP] appointment: default type for staff
Currently creating a new appointment for
a user-based appointment type from gantt
does not populate the appointment type nor
the title.

We add a context key to the create call from gantt
so that we can more freely assign default values
when creating appointments, which is implicit when
using this gantt view.

task-3434098

X-original-commit: bd747d758a77c116d370418128b5e35fde2e7fc5
Part-of: odoo/enterprise#49295