[FIX] stock_barcode: hide barcode setting for picking types
Currently the Barcode App tab in picking.type show for all picking
types, but it doesn't make sense for the newly added Dropship picking
type code (as of saas-16.2 odoo/odoo#109511, before that we have no way
of hiding it for this picking type) and any other picking types that
aren't accessible in the barcode app (i.e. as of saas-16.4 for Repair
picking type odoo/odoo#106911 and manufacturing until saas-16.4)

Therefore let's hide it. Note that additional logic is needed in
saas-16.4 onwards so that the setting shows for mrp (i.e. when mrp
barcode feature was added)

closes odoo/enterprise#45568

Task: 3441559
Signed-off-by: Steve Van Essche <svs@odoo.com>