[FIX] industry_fsm_sale: use ordered qty for fixed price service
The sign report on a fsm task for a prepaid service does not represent the true amount and hours. Because the service is prepaid/fixed price, it should make calculations according to ordered qty, not the delivered qty.

Streps to reproduce:
1. Configure a product type service as prepaid/fixed price to create a task in a field service project 
2. Add the product to a SO with quantity 1 and price 100 
3. Confirm the SO
4. Go to the task and timesheet 3 hours
5. Click on Sign report

Current Behavior:
The report will show the amount 300 instead of 100.

Expected Behavior:
The amount should be 100, as calculated by the ordered qty.

OPW-3449841

closes odoo/enterprise#46423

X-original-commit: 5edc7dd72fa535dc5cb79844431dc0bb0380d06c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Hamza Islam (hisl) <hisl@odoo.com>