[FIX] account_consolidation: copy chart with all groups and accounts
Steps to reproduce:
- Go to Consolidation / Configuration / Consolidation
- Create a new chart (e.g. Chart X)
- For Chart X, create the following groups and subgroups:
  * Group A
  * Group AA (Parent: Group A)
  * Group AAA (Parent: Group AA)
- For Chart X, create the following accounts:
  * Account A (Group: Group AAA)
  * Account B (Group: None)
- Duplicate Chart X
The copied chart is correctly linked to the copied root group (i.e. Group A (copy)) and
its direct child (i.e. Group AA (copy)), but subgroups on a deeper level (i.e. Group AAA(copy))
is linked to the original chart.
Also, only accounts linked to a group is copied (i.e. Account A).

Solution:
Recursively set the chart of copied groups and subgroups to the id of the copied chart
and include accounts with no group in the copy.

opw-3381529

closes odoo/enterprise#45778

X-original-commit: aa5f73b0bb172240c977ad160b3966234def0f35
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>