[IMP] *_spreadsheet_*: Update o_spreadsheet library to version 16.5.0-alpha.3
[IMP] spreadsheet: change notify behavior

This commit improves the notifications that can be sent
to the user interface in three ways

1 - change responsibilities

Spreadsheet has been developed so that it isn't possible
to notify several times with the same message.
We believe that this feature should not be general but
specific to each notifying element.
Technically, this means no longer using "tag" on the
notifyUser payload

2 - Have different notification categories.

The spreadsheet integration currently allows having only
one level of notification. We would like to have several
such as:
- danger
- warning
- info
- success

3 - Notify user with sticky and no sticky notifications

odoo task id 3414653

closes odoo/enterprise#45558

Related: odoo/odoo#131407
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>