[IMP] web_(grid|map): these views are *not* mobile-friendly
A new, deterministic mechanism to decide which view should be displayed
in a mobile environment for window actions has been created.

This allows us to remove the 'mobile-friendly' flag for grid and map
views which were added to override the kanban views in some cases.

The modules that use these views (timesheet_grid and industry_fsm) have
been modified so that the proper mobile view is set on the window action
where needed.

Task-3460374

closes odoo/enterprise#46562

Related: odoo/odoo#133608
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>