[FIX] l10n_mx_edi: getter on a None value
In the l10n_mx_edi module, xpaths are done on the
report_payment_receipt_document template of the account module.
In one of those xpaths, a t-set is done using a value from cfdi.
cfdi is coming from the "values" variable but can be None.
A check of the existence of cfdi is always done before its use in the
other xpaths of the module. This commit adds this check to the
problematic xpath.

closes odoo/enterprise#45628

X-original-commit: 8c603bf3c0452ff0116dc380b4a10c947d20d6d0
Signed-off-by: Laurent Smet (las) <las@odoo.com>