[FIX] appointment: prevent multiple bookings per resource
Before this change, we considered that a resource booked in
an appointment type could be booked in another at the same
time.

Now, we consider all bookings - of other appointment types too -
when evaluating whether a resource is available.

A test is added to ensure this bahavior.

Task-3460526

X-original-commit: a11bd3b5059a1125922c1aefe55a7350f2a61968
Part-of: odoo/enterprise#47059