[IMP] industry_fsm: improve navigate button in fsm task form.
In this commit, the functionality of the 'Navigate' button in the FSM task form
has been improved. The 'Navigate to' button is now enhanced to better interact
with the task. Specifically, the button is hidden when no partner is assigned to
the task. This enhancement prevents the 'Navigate to' button from causing an
'invalid field' error when clicked, as it's not applicable in such cases.

closes odoo/enterprise#45783

Task: 3460000
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>