[FIX] sale_amazon: update log  when  proxy connection not established
This issue occurs when the user tries to connect to sale amazon and at that time
, the user gets problems due to network problems, proxy server configuration
errors, etc.

Message-Could not establish the connection to the proxy.

The logger is updated to use the 'warning' level instead of the 'error' level.
This change reflects a less severe logging level for cases when timeout while\
 connection of sale amazon

sentry-4366899402

closes odoo/enterprise#45671

X-original-commit: 8398b847590baf1d329b172c0d8a64d1d4b0115d
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>