[REF] knowledge: render embedded views/links in frontend
This commit allows to do all the rendering for Embedded views in frontend,
without relying on an rpc call and backend rendering. This cleans up the
knowledge_article model and centralizes templates used as Behavior blueprints in
`knowledge_editor.xml`

Since [1], favorites for embedded views are saved in the HTML arch of the body,
and do not create real `ir.filters` records. This means that we can now use xml
ids for embedded views without interfering on the original view in its original
module. XmlId of the action window now becomes the default stored information
to refer to it in an embed, but act_window objects are still supported.

[1]: https://github.com/odoo/enterprise/commit/cc7afb66f09e11329fc2721d8bf33e2de66b1a36

Task-3284243

closes odoo/enterprise#41648

Related: odoo/upgrade#4758
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>