[FIX] knowledge: fix missing display name for users in kebab menu
When loading a many-to-one field that is invisible in the arch of the
view, the system no longer fetches the display name of those records
for performance reasons. It now only fetches the id of the record.

In Knowledge, there is an issue induced by that change: The system no
longer load (1) the name of the user who create the article and (2) the
name of the last user who edit the article. As a result, those values
are no longer display in the kebab menu of Knowledge.

This commit will fix that issue by adding new field dependencies in the
TopBar component. The system will now load the display name as it was the
case before.

task-3452142

closes odoo/enterprise#45077

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>