[IMP] account_{reports, consolidation, cash_basis},l10n_be_reports: pdf export
During the html and css change (https://github
.com/odoo/enterprise/commit/2d774344109c2f9b55d9ea6160ae99aeca4f0420), the PDF
were not modified.

We modified the pdf_export_main template to remove stuff that was not necessary
in the report like buttons, ellipsis, etc. And entirely reworked the css the
report to match the one from the interface.

Task: 3437257
Part-of: odoo/enterprise#45792